Android 15.0.0 Release 20 (BP1A.250305.019)

# -----BEGIN PGP SIGNATURE-----
#
# iF0EABECAB0WIQRDQNE1cO+UXoOBCWTorT+BmrEOeAUCZ8epZAAKCRDorT+BmrEO
# eBGSAJsFh8d8gfdI4zxAc4KibqtETT4nLQCcDrm7A1qCjUkXcRE6DSMAz19CUXo=
# =F+JD
# -----END PGP SIGNATURE-----
# gpg: Signature made Wed Mar  5 03:31:16 2025 EET
# gpg:                using DSA key 4340D13570EF945E83810964E8AD3F819AB10E78
# gpg: Good signature from "The Android Open Source Project <initial-contribution@android.com>" [ultimate]

* tag 'android-15.0.0_r20': (85 commits)
  Cleared persistent wallpaper repo value after reading it
  Create utility class for testing android.app.WallpaperInfo with Robolectric
  Ensure that wallpaper attribution entries are not null
  Simplify adding color update flows
  Revert "Fix the layout of the grid option customization (1/2)"
  Import translations. DO NOT MERGE ANYWHERE
  Migrates Monet's Style Enum to @IntDef
  Fix the layout of the grid option customization (1/2)
  Fix foldable set wallpaper dialog content description
  Bind Toolbar colors (1/2)
  Don't show set wallpaper dialog after screen rotation
  Fix not showing low res wallpaper in small preview
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Make preview smartspace corresponds to the selected clock (2/3)
  Update layout for foldable apply wallpaper screen
  Expose sysui runtime values to wallpaper picker (1/2)
  Import translations. DO NOT MERGE ANYWHERE
  Cleanup preview wallpaper model before launching preview again
  Fix always land on lock screen preview
  ...

+++ src/com/android/wallpaper/picker/preview/ui/WallpaperPreviewActivity.kt
@@ -98,19 +98,27 @@ class WallpaperPreviewActivity :
             refreshCreativeCategories = intent.getBooleanExtra(SHOULD_CATEGORY_REFRESH, false)
         }
 
-        val wallpaper: WallpaperModel? =
+        val wallpaper: WallpaperModel =
             if (isNewPickerUi || isCategoriesRefactorEnabled) {
-                persistentWallpaperModelRepository.wallpaperModel.value
-                    ?: intent
-                        .getParcelableExtra(EXTRA_WALLPAPER_INFO, WallpaperInfo::class.java)
-                        ?.convertToWallpaperModel()
+                val model =
+                    if (savedInstanceState != null) {
+                        wallpaperPreviewViewModel.wallpaper.value
+                    } else {
+                        persistentWallpaperModelRepository.wallpaperModel.value
+                            ?: intent
+                                .getParcelableExtra(EXTRA_WALLPAPER_INFO, WallpaperInfo::class.java)
+                                ?.convertToWallpaperModel()
+                    }
+                persistentWallpaperModelRepository.cleanup()
+                model
             } else {
                 intent
                     .getParcelableExtra(EXTRA_WALLPAPER_INFO, WallpaperInfo::class.java)
                     ?.convertToWallpaperModel()
-            }
-
-        wallpaper ?: throw UnsupportedOperationException()
+            } ?: throw IllegalStateException("No wallpaper for previewing")
+        if (savedInstanceState == null) {
+            wallpaperPreviewRepository.setWallpaperModel(wallpaper)
+        }
 
         val navController =
             (supportFragmentManager.findFragmentById(R.id.wallpaper_preview_nav_host)
@@ -132,9 +140,6 @@ class WallpaperPreviewActivity :
         WindowCompat.setDecorFitsSystemWindows(window, ActivityUtils.isSUWMode(this))
         val isAssetIdPresent = intent.getBooleanExtra(IS_ASSET_ID_PRESENT, false)
         wallpaperPreviewViewModel.isNewTask = intent.getBooleanExtra(IS_NEW_TASK, false)
-        if (savedInstanceState == null) {
-            wallpaperPreviewRepository.setWallpaperModel(wallpaper)
-        }
         val whichPreview =
             if (isAssetIdPresent) WallpaperConnection.WhichPreview.EDIT_NON_CURRENT
             else WallpaperConnection.WhichPreview.EDIT_CURRENT
@@ -150,7 +155,7 @@ class WallpaperPreviewActivity :
             liveWallpaperDownloader.initiateDownloadableService(
                 this,
                 wallpaper,
-                registerForActivityResult(ActivityResultContracts.StartIntentSenderForResult()) {}
+                registerForActivityResult(ActivityResultContracts.StartIntentSenderForResult()) {},
             )
         }
 
@@ -158,7 +163,9 @@ class WallpaperPreviewActivity :
             (wallpaper as? WallpaperModel.LiveWallpaperModel)
                 ?.creativeWallpaperData
                 ?.creativeWallpaperEffectsData
-        if (creativeWallpaperEffectData != null) {
+        if (
+            creativeWallpaperEffectData != null && !creativeEffectsRepository.isEffectInitialized()
+        ) {
             lifecycleScope.launch {
                 creativeEffectsRepository.initializeEffect(creativeWallpaperEffectData)
             }
@@ -195,15 +202,27 @@ class WallpaperPreviewActivity :
         }
     }
 
-    override fun onDestroy() {
+    override fun onPause() {
+        super.onPause()
+
+        // When back to main screen user could launch preview again before it's fully destroyed and
+        // it could clean up the repo set by the new launching call, move it earlier to on pause.
         if (isFinishing) {
             persistentWallpaperModelRepository.cleanup()
+        }
+    }
+
+    override fun onDestroy() {
+        if (isFinishing) {
             // ImageEffectsRepositoryImpl is Activity-Retained Scoped, and its injected
             // EffectsController is Singleton scoped. Therefore, persist state on config change
             // restart, and only destroy when activity is finishing.
             imageEffectsRepository.destroy()
+            // CreativeEffectsRepository is Activity-Retained Scoped, and its injected
+            // EffectsController is Singleton scoped. Therefore, persist state on config change
+            // restart, and only destroy when activity is finishing.
+            creativeEffectsRepository.destroy()
         }
-        creativeEffectsRepository.destroy()
         liveWallpaperDownloader.cleanup()
         // TODO(b/333879532): Only disconnect when leaving the Activity without introducing black
         //  preview. If onDestroy is caused by an orientation change, we should keep the connection
@@ -268,7 +287,7 @@ class WallpaperPreviewActivity :
             isAssetIdPresent: Boolean,
             isViewAsHome: Boolean = false,
             isNewTask: Boolean = false,
-            shouldCategoryRefresh: Boolean
+            shouldCategoryRefresh: Boolean,
         ): Intent {
             val isNewPickerUi = BaseFlags.get().isNewPickerUi()
             val isCategoriesRefactorEnabled =
@@ -329,7 +348,7 @@ class WallpaperPreviewActivity :
             isAssetIdPresent: Boolean,
             isViewAsHome: Boolean = false,
             isNewTask: Boolean = false,
-            shouldRefreshCategory: Boolean
+            shouldRefreshCategory: Boolean,
         ): Intent {
             val intent = Intent(context.applicationContext, WallpaperPreviewActivity::class.java)
             if (isNewTask) {
@@ -367,7 +386,7 @@ class WallpaperPreviewActivity :
                     ImageWallpaperInfo(data),
                     isAssetIdPresent,
                     isViewAsHome,
-                    isNewTask
+                    isNewTask,
                 )
             // Both these lines are required for permission propagation
             intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
